Añadida seguridad al userbooks , mejorado boton de borrar, mostrado dinamico si no tienes libros y recortada descripcion en le home

+++ src/views/UserBooks.vue
@@ -1,29 +1,41 @@
 <template>
   <div style="background:rgb(230,230,230)">
-    <NavBar/>
+    <NavBar />
     <div class="container" style="margin-top:100px;">
-      <div class="row" v-for="libro in libros" v-bind:key="libro.titulo" style="background-color:white;margin-bottom:30px;border-radius:7px;">
-    <div class="col d-flex d-xxl-flex justify-content-center justify-content-xxl-center align-items-xxl-center" style="padding: 7px;">
-      <img src="https://www.gtm-store.com/wp-content/uploads/2021/10/Presentacion-05.jpg" style="width: 100%;border-radius: 8px;" />
-    </div>
-    <div class="col-8" style="padding-top: 13px;">
-        <h1 style="text-align:left;">{{libro.titulo}}</h1>
-        <hr style="margin-top: 5px;" />
-        <p class="text-break" style="max-height: 96px;text-align:left;">{{libro.descripcion}}</p>
-        <div style="padding-top: 5px;padding-bottom: 5px;text-align:left;">
-          <button class="btn btn-primary" v-on:click="ver(libro.idlibro)" type="button" style="margin-right: 12px;width: 70px;padding: 5px;color: var(--bs-white);background: var(--bs-teal);border-color: var(--bs-white);">Ver
-            <i class="fa fa-eye" style="padding-left: 5px;"></i>
-          </button>
-          <button class="btn btn-primary" type="button" style="width: 80px;padding: 5px;background: var(--bs-teal);border-color: rgb(255, 255, 255);border-top-color: rgb(255,;border-right-color: 255,;border-bottom-color: 255);border-left-color: 255,;">Editar
-            <i  class="fa fa-edit" style="padding-left: 5px;"></i>
-          </button>
+      <div class="row" v-for="libro in libros" v-bind:key="libro.titulo"
+        style="background-color:white;margin-bottom:30px;border-radius:7px;">
+        <div class="col d-flex d-xxl-flex justify-content-center justify-content-xxl-center align-items-xxl-center"
+          style="padding: 7px;">
+          <img src="https://www.gtm-store.com/wp-content/uploads/2021/10/Presentacion-05.jpg"
+            style="width: 100%;border-radius: 8px;" />
         </div>
+        <div class="col-8" style="padding-top: 13px;">
+          <h1 style="text-align:left;">{{libro.titulo}}</h1>
+          <hr style="margin-top: 5px;" />
+          <p class="text-break" style="max-height: 96px;text-align:left;">{{libro.descripcion}}</p>
+          <div style="padding-top: 5px;padding-bottom: 5px;text-align:left;">
+            <button class="btn btn-primary" v-on:click="verlibro(libro.idlibro)" type="button"
+              style="margin-right: 12px;width: 70px;padding: 5px;color: var(--bs-white);background: var(--bs-teal);border-color: var(--bs-white);">Ver
+              <i class="fa fa-eye" style="padding-left: 5px;"></i>
+            </button>
+            <button class="btn btn-primary" type="button"
+              style="width: 80px;padding: 5px;background: var(--bs-teal);border-color: rgb(255, 255, 255);border-top-color: 255;border-right-color: 255,;border-bottom-color: 255;border-left-color: 255,;">Editar
+              <i class="fa fa-edit" style="padding-left: 5px;"></i>
+            </button>
+          </div>
+        </div>
+        <div class="col-1 text-end" style="padding: 21px;">
+          <i id="borrar" v-on:click="borrar(libro.idlibro)" class="fa-solid fa-trash-can"
+            style="font-size: 25px;color: var(--bs-red);"></i>
+        </div>
+      </div>
     </div>
-     <div class="col-1 text-end" style="padding: 21px;"><i v-on:click="borrar(libro.idlibro)" class="fa-solid fa-trash-can" style="font-size: 25px;color: var(--bs-red);"></i></div>
-</div>
+
+    <div v-if="this.libros.lenght == 0">
+      <h1>HOLAAAAAAAAAAAAA</h1>
     </div>
-  
-    <Footer/>
+
+    <Footer />
   </div>
 
 </template>
@@ -35,6 +47,7 @@
 import NavBar from "../components/NavBar.vue";
 import Footer from "../components/Footer.vue";
 import env from "../environment.js";
+import Swal from 'sweetalert2';
 
 
 
@@ -48,39 +61,69 @@ export default {
   },
   data(){
     return{
-     userid: String,
+     userid: '',
      libros: {},
+     session: {},
+     idlibroprueba: '',
     }
   },
   async created (){
-     this.userid = await this.$route.params.userid;
-     console.log(this.userid);
-
+    this.userid = await this.$route.params.userid;
+    this.session = await env.supabase.auth.session();
+    if(!this.session){
+      this.$router.push("/home");
+    }
+     
     const { data: Libro, error} = await env.supabase
     .from('Libro')
     .select("*")
     .eq('iduser',this.userid)
  
+   
+    this.idlibroprueba = Libro[0].iduser;
+    if (this.idlibroprueba != this.userid) {
+      this.$router.push("/home");
+    }
+
+
+    
     if(error){
       console.log(error);
       this.$router.push("/home");
-      }else{
-        this.libros = Libro;
-      }
+    }else{
+      this.libros = Libro;
+    }
     
     
 
  },
  methods:{
    async borrar(id){ 
-    await env.supabase
+
+  Swal.fire({
+  title: 'Estas seguro de que desea borrar este libro?',
+  icon: 'warning',
+  showCancelButton: true,
+  confirmButtonColor: '#c91e1e',
+  color:'#d33',
+  iconColor:'#c91e1e',
+  cancelButtonColor: '#8f9491',
+  confirmButtonText: 'Borrar'
+  }).then((result) => {
+  if (result.isConfirmed) {
+     env.supabase
     .from('Libro')
     .delete()
     .eq('idlibro', id)
-   },
-    ver(id){ 
-    this.$router.push('/book/' + id)
-   }
+
+    this.$router.go();
+  }
+})
+},
+verlibro(id){ 
+  this.$router.push('/book/' + id)
+},
+  
  }
   
 };
@@ -88,4 +131,7 @@ export default {
 
 <style scoped>
 @import url('https://fonts.googleapis.com/css2?family=Amatic+SC&display=swap');
+#borrar:hover {
+  cursor: pointer;
+}
 </style>
